포탄 서버 연동 테스트

+++ Project/SimpleGame/Input.cpp
@@ -93,8 +93,8 @@ void SpecialKeyUpInput(int key, int x, int y) {
 void MouseInput(int button, int state, int x, int y) {
 	if (button == GLUT_LEFT_BUTTON && state == GLUT_DOWN && mouseFlag == false) {
 		mouseFlag = true;
-		D_INPUT->mx = (float)x;
-		D_INPUT->my = (float)y;
+		D_INPUT->mx = x;
+		D_INPUT->my = y;
 	}
 	else if (state == GLUT_UP)
 		mouseFlag = false;
