+++ Project/SimpleGame/Input.h
@@ -11,7 +11,7 @@ public:
 	bool isMouseDown();
 	void setDeltaTime(float t);
 	float getDeltaTime();
-	float mx, my;
+	short mx, my;
 private:
 	float deltaTime;
 };
